% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentationMC_support.R
\docType{package}
\name{AugmentationMC}
\alias{AugmentationMC}
\title{Monte Carlo based tools and algorithms for Data Augmentation}
\description{
This library provides a set of functions for the implementation of Monte Carlo
based algorithms for the augmentation of both qualitative and quantitative data.
This includes a series of Monte Carlo simulation functions, as well as an implementation
of a Metropolis-Hastings variant of the Markov Chain Monte Carlo algorithm, with
a series of functions for the evaluation of chains.
}
\author{
Lloyd A. Courtenay
}
