% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\name{autocorrelation}
\alias{autocorrelation}
\title{Autocorrelation Function (ACF) Plots}
\usage{
autocorrelation(trace, dimension = NULL, main = NULL)
}
\arguments{
\item{trace}{an MCMC_Trace object}

\item{dimension}{an integer defining which dimension of the distribution should
be visualised (if the trace is multivariate)}

\item{main}{A string defining the title for the plot. If no main is defined then
"Correlogram" will be used as default.}
}
\description{
Computation and plotting of a Autocorrelation Function plot (or Correlogram)
}
\details{
ACF Plots can be used in combination with trace plots to visualise
and measure the serial correlation of a chain. Effective use of MCMC
should see the curve approximate 0 quickly, while chains that take a while to
reach 0 indicate the step size of MCMC may not be ideal for this particular
case study. Likewise, another common issue in MCMCs may be that the size of the trace
is too small, and increasing the number of iterations will improve results.
From this perspective, the user can assess how well and how
quickly the MCMC chain has converged on the actual distribution.
}
\examples{

data("single_moon")

trial_1 <- MCMC(as.matrix(single_moon[,1]), 10000, step_size = 1)
trial_2 <- MCMC(as.matrix(single_moon[,1]), 10000, step_size = 0.5)
trial_3 <- MCMC(as.matrix(single_moon[,1]), 10000, step_size = 0.1)


# example of convergence
autocorrelation(trial_1, main = "Step Size = 1")

# example of poor convergence
autocorrelation(trial_2, main = "Step Size = 0.5")

# example of no convergence
autocorrelation(trial_3, main = "Step Size = 0.1")

#

}
\seealso{
\code{\link{MCMC}}, \code{\link{MCMC_Trace}}, \code{\link{burn_in}},
\code{\link{sample_from_trace}},
\code{\link{effective_sample_size}}, \code{\link{TOST}},
\code{\link{trace_plot}}
}
\author{
Lloyd A. Courtenay
}
