% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_multimodalMC.R
\name{bivariate_multimodalMC}
\alias{bivariate_multimodalMC}
\title{Monte Carlo Simulation of Bivariate Multimodal Data}
\usage{
bivariate_multimodalMC(data_set, n_simulations, num_breaks = 0)
}
\arguments{
\item{data_set}{An object of type data.frame containing a single categorical
variable and a single numerical variable that the user wishes to augment}

\item{n_simulations}{An integer defining the number of simulations}

\item{num_breaks}{An optional parameter that can be used to fine-tune the calculation
of the probability distribution from which to sample from.}
}
\value{
simulated data
}
\description{
The present function can be used to simulate a set of 2 variables; 1 categorical
and 1 numerical.
}
\section{Details}{

This function uses the statistical chain rule to calculate probability
vectors from which to sample from so as to sample from the multivariate
categorical-numerical distribution
}

\examples{

bivariate_trial <- data.frame(x = c(rnorm(100, 0, 2), rnorm(100, -4, 3)),
                              sample = as.factor(c(rep("Group_1", 100),
                                                   rep("Group_2", 100))))

augmented_trial <- bivariate_multimodalMC(bivariate_trial, 100)

par(mfrow = c(2,2))
hist(bivariate_trial[bivariate_trial$sample == "Group_1",1],
     main = "Original Data Group 1",
     xlab = "Group 1")
hist(c(bivariate_trial[bivariate_trial$sample == "Group_1",1],
       augmented_trial[augmented_trial$sample == "Group_1",1]),
     main = "Augmented Data Group 1",
     xlab = "Group 1")
hist(bivariate_trial[bivariate_trial$sample == "Group_2",1],
     main = "Original Data Group 2",
     xlab = "Group 2")
hist(c(bivariate_trial[bivariate_trial$sample == "Group_2",1],
       augmented_trial[augmented_trial$sample == "Group_2",1]),
     main = "Augmented Data Group 2",
     xlab = "Group 2")
par(mfrow = c(1,1))

#

}
\author{
Lloyd A. Courtenay
}
