% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\name{burn_in}
\alias{burn_in}
\title{Burn-In Sample}
\usage{
burn_in(traceObject, burn_in = 0.25)
}
\arguments{
\item{traceObject}{an MCMC_Trace object}

\item{burn_in}{a numeric value between 0 and 1 defining the percentage of burn-in
to be removed. Default is 0.25}
}
\value{
An MCMC_Trace object excluding the burn-in period
}
\description{
The present function can be used to remove the burn in from a MCMC Trace Object
}
\examples{

data("single_moon")

augmentation_trial <- MCMC(as.matrix(single_moon), 10000, step_size = 1)

augmentation_trial <- burn_in(augmentation_trial, 0.25)

#

}
\author{
Lloyd A. Courtenay
}
