% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalMC.R
\name{categoricalMC}
\alias{categoricalMC}
\title{Monte Carlo Simulations for Categorical Variables}
\usage{
categoricalMC(data_set, n_simulations)
}
\arguments{
\item{data_set}{An object of type data.frame containing either uni- or multivariate
categorical variables}

\item{n_simulations}{An integer defining the number of simulations}
}
\value{
simulated data
}
\description{
The present function can be used to simulate categorical (qualitative) data
}
\section{Details}{

This function uses the statistical chain rule to calculate multivariate probability
vectors associated with each categorical variable from which to sample from
}

\examples{

data(multiclass_dataset)
data_set <- multiclass_dataset[,c(3, 4)]

# augment data

augmented_data <- categoricalMC(data_set, 1000)

# visualise data

par(mfrow = c(2,2))
barplot(table(data_set$Sample_1),
        main = "Original Sample 1")
barplot(table(augmented_data$Sample_1),
        main = "Augmented Sample 1")
barplot(table(data_set$Sample_2),
        main = "Original Sample 2")
barplot(table(augmented_data$Sample_2),
        main = "Augmented Sample 2")
par(mfrow = c(1,1))

# statistical proportion test

n_original <- nrow(data_set)
n_augmented <- nrow(augmented_data)

# example A1

sample_1_original_A1 <- table(data_set$Sample_1)[1]
sample_1_augmented_A1 <- table(augmented_data$Sample_1)[1]

prop.test(x = c(sample_1_original_A1,
                sample_1_augmented_A1),
          n = c(n_original, n_augmented))

# example B1

sample_1_original_B1 <- table(data_set$Sample_1)[2]
sample_1_augmented_B1 <- table(augmented_data$Sample_1)[2]

prop.test(x = c(sample_1_original_B1,
                sample_1_augmented_B1),
          n = c(n_original, n_augmented))

# example C1

sample_1_original_C1 <- table(data_set$Sample_1)[3]
sample_1_augmented_C1 <- table(augmented_data$Sample_1)[3]

prop.test(x = c(sample_1_original_C1,
                sample_1_augmented_C1),
          n = c(n_original, n_augmented))

# example A2

sample_2_original_A2 <- table(data_set$Sample_2)[1]
sample_2_augmented_A2 <- table(augmented_data$Sample_2)[1]

prop.test(x = c(sample_2_original_A2,
                sample_2_augmented_A2),
          n = c(n_original, n_augmented))

# example B2

sample_2_original_B2 <- table(data_set$Sample_2)[2]
sample_2_augmented_B2 <- table(augmented_data$Sample_2)[2]

prop.test(x = c(sample_2_original_B2,
                sample_2_augmented_B2),
          n = c(n_original, n_augmented))

#

}
\author{
Lloyd A. Courtenay
}
