% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentationMC_support.R
\docType{data}
\name{circles}
\alias{circles}
\title{Toy Circles Dataset}
\format{
A data frame with 500 rows and 2 variables (x, y)
}
\description{
A toy dataset containing x and y coordinates for a multivariate inverse-gaussian distribution
}
\author{
Lloyd A. Courtenay
}
\keyword{datasets}
