% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalMC.R
\name{conditionalMC}
\alias{conditionalMC}
\title{Conditional Monte Carlo Simulations for Categorical and Numeric Variables}
\usage{
conditionalMC(data_set, n_simulations, num_breaks = 0)
}
\arguments{
\item{data_set}{An object of type data.frame containing either a single categorical
variable associated with a series of numeric variables, or a single numeric variable associated
with a series of categorical variables.}

\item{n_simulations}{An integer defining the number of simulations}

\item{num_breaks}{An optional parameter that can be used to fine-tune the calculation
of the probability distribution from which to sample from.}
}
\value{
simulated data
}
\description{
The present function can be used to simulate a set of numeric values related to
a qualitative variable, or a set of categorical values related to a quantitative variable
}
\section{Details}{

This function uses the statistical chain rule to calculate multivariate probability
vectors from which to sample from, firstly calculating the probability of
the conditioning variable, and then sampling values associated with these
conditioning variables.
}

\examples{

# example 1 ---------------------

data("multiclass_dataset")
trial <- conditionalMC(multiclass_dataset[,c(1,2,3)], 100)

plot(multiclass_dataset[multiclass_dataset$Sample_1 == "A1",1],
     multiclass_dataset[multiclass_dataset$Sample_1 == "A1",2],
     asp = 1, pch = 1,
     xlim = c(min(multiclass_dataset$x), max(multiclass_dataset$x)),
     ylim = c(min(multiclass_dataset$y), max(multiclass_dataset$y)))
points(multiclass_dataset[multiclass_dataset$Sample_1 == "B1",1],
       multiclass_dataset[multiclass_dataset$Sample_1 == "B1",2],
       pch = 1, col = "red")
points(multiclass_dataset[multiclass_dataset$Sample_1 == "C1",1],
       multiclass_dataset[multiclass_dataset$Sample_1 == "C1",2],
       pch = 1, col = "blue")

points(trial[trial$Sample_1 == "A1", 1],
       trial[trial$Sample_1 == "A1", 2],
       pch = 19, col = "black", cex = 2)
points(trial[trial$Sample_1 == "B1", 1],
       trial[trial$Sample_1 == "B1", 2],
       pch = 19, col = "red", cex = 2)
points(trial[trial$Sample_1 == "C1", 1],
       trial[trial$Sample_1 == "C1", 2],
       pch = 19, col = "blue", cex = 2)

# example 2 ---------------------

# generate toy data

x1 <- rnorm(100, mean = 2, sd = 1)
y1 <- rnorm(100, mean = 2, sd = 1)
x2 <- rnorm(100, mean = -2, sd = 1)
y2 <- rnorm(100, mean = -2, sd = 1)
group_1 <- cbind(x1, y1, rep("Group_1", 100))
group_2 <- cbind(x2, y2, rep("Group_2", 100))
samples <- data.frame(rbind(group_1, group_2))
samples[,1] <- as.numeric(samples[,1])
samples[,2] <- as.numeric(samples[,2])
samples[,3] <- as.factor(samples[,3])

# augment

trial2 <- conditionalMC(samples, 100, num_breaks = 20)

# plot

plot(samples[samples$V3 == "Group_1", 1], samples[samples$V3 == "Group_1", 2],
     asp = 1, xlim = c(-5, 5), ylim = c(-5, 5), pch = 1)
points(samples[samples$V3 == "Group_2", 1], samples[samples$V3 == "Group_2", 2],
       col = "red", pch = 1)

points(trial2[trial2$V3 == "Group_1", 1], trial2[trial2$V3 == "Group_1", 2],
       pch = 19)
points(trial2[trial2$V3 == "Group_2", 1], trial2[trial2$V3 == "Group_2", 2],
       pch = 19, col = "red")

#

}
\author{
Lloyd A. Courtenay
}
