% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{Markov Chain Monte Carlo for Multivariate Numeric Distributions}
\usage{
MCMC(data_set, n_iterations, step_size = 0.5, num_breaks = 0)
}
\arguments{
\item{data_set}{An matrix or array of numeric data}

\item{n_iterations}{The number of iterations to compute of the Markov Chain}

\item{step_size}{The step size of the Markov Chain. 0.5 is the default step size.}

\item{num_breaks}{An optional parameter that can be used to fine-tune the calculation
of the probability distribution from which to sample from.}
}
\value{
Markov Chain Monte Carlo Trace S4 Object
}
\description{
The present function can be used to simulate numerical data using the
Metropolis-Hastings 1st Order Variant of the
Markov Chain Monte Carlo algorithm.
}
\section{Details}{

This function uses the the Metropolis-Hastings 1st Order Variant (Metropolis et al., 1953;
Hastings, 1970) of the Markov Chain Monte Carlo (MCMC) algorithm (Gamerman and Lopes, 2006),
to simulate numerical data from a multivariate distribution.

MCMCs are stochastic algorithms that inherit from Monte Carlo based methods
by generating random values from a given probability distribution, while using
Markovian chains to explore the distribution via "random walks". The Metropolis-Hastings
algorithm is used to define an acceptance criterion for a proposed state of the Markov
Chain.

State probabilities for this funciton are calculated using the chain rule in statistics.
Probability density functions for each dimension's distribution were additionally smoothed
using polynomial interpolation.
}

\section{Bibliography}{


Gamerman, D.; Lopes, H.R. (2006) Markov Chain Monte Carlo, Boca Raton: Chapman & Hall.

Hastings, W.K. (1970) Monte-Carlo sampling methods using Markov Chains and their applications, Biometrika, 57:97-109

Metropolis, N.; Rosenbluth, A.W.; Rosenbluth, M.N.; Teller, A.H.; Teller, E. (1953)
Equation of state calculations by fast computing machines, Journal of Chemical Physics.
21:1087-1092
}

\examples{

data("single_moon")

augmentation_trial <- MCMC(as.matrix(single_moon), 10000, step_size = 1)

trace <- sample_from_trace(augmentation_trial, 100)

plot(single_moon, asp = 1, pch = 19)
points(trace[,1], trace[,2], col = "red", pch  = 19)

#

}
\seealso{
evaluate_single_trace...
}
\author{
Lloyd A. Courtenay
}
