% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\docType{class}
\name{MCMC_Trace-class}
\alias{MCMC_Trace-class}
\title{An S4 Class for Markov Chain Monte Carlo traces}
\description{
An S4 Class for Markov Chain Monte Carlo traces
}
\section{Slots}{

\describe{
\item{\code{Trace}}{A numerical matrix containing the MCMC trace}

\item{\code{burn_in}}{A boolean value indicating whether the burn in period of the
trace has been removed}
}}

\seealso{
\code{\link{MCMC}}
}
