% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentationMC_support.R
\docType{data}
\name{moons}
\alias{moons}
\title{Toy Moons Dataset}
\format{
A data frame with 1000 rows and 2 variables (x, y)
}
\description{
A toy dataset containing x and y coordinates of two multivariate moons
}
\author{
Lloyd A. Courtenay
}
\keyword{datasets}
