% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentationMC_support.R
\docType{data}
\name{moons_qual_quant}
\alias{moons_qual_quant}
\title{Toy Qualitative and Quantitative Moons Dataset}
\format{
A data frame with 1000 rows and 3 variables; 2 quantitative (x, y), 1 qualitative
(category)
}
\description{
A toy dataset containing x and y coordinates of two multivariate moons with an additional factor column
}
\author{
Lloyd A. Courtenay
}
\keyword{datasets}
