% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentationMC_support.R
\docType{data}
\name{multiclass_dataset}
\alias{multiclass_dataset}
\title{Toy Complex and Imbalanced Qualitative and Quantitative Dataset}
\format{
A data frame with 210 rows and 4 variables; 2 quantitative (x, y), 2 qualitative (Sample_1, Sample_2).
Sample_1 consistes of three classes, A1, B1, and C1, of which the latter is imbalanced (number of individuals = 10)
}
\description{
A toy dataset containing x and y coordinates of two multivariate distributions with
2 seperate categorical classes
}
\author{
Lloyd A. Courtenay
}
\keyword{datasets}
