% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodalMC.R
\name{multimodalMC}
\alias{multimodalMC}
\title{Monte Carlo Simulation of Multivariate Multimodal Data}
\usage{
multimodalMC(data_set, n_simulations, num_breaks = 0)
}
\arguments{
\item{data_set}{An object of type data.frame containing both qualitative
and quantitative variables that the user wishes to augment}

\item{n_simulations}{An integer defining the number of simulations}

\item{num_breaks}{An optional parameter that can be used to fine-tune the calculation
of the probability distribution from which to sample from.}
}
\value{
simulated data
}
\description{
The present function can be used to simulate a mixture of both categorical (qualitative)
and numeric (quantitative) variables using a Monte Carlo approach
}
\section{Details}{

This function uses the statistical chain rule to calculate multivariate probability
vectors from which to sample from, firstly calculating the probability of
qualitative variables, and then sampling numeric values associated with these
qualitative variables.
}

\examples{

data(multiclass_dataset)

trial_simulation <- multimodalMC(multiclass_dataset, n_simulations = 100)

plot(multiclass_dataset[multiclass_dataset$Sample_1 == "A1",1],
     multiclass_dataset[multiclass_dataset$Sample_1 == "A1",2],
     asp = 1, pch = 1,
     xlim = c(min(multiclass_dataset$x), max(multiclass_dataset$x)),
     ylim = c(min(multiclass_dataset$y), max(multiclass_dataset$y)))
points(multiclass_dataset[multiclass_dataset$Sample_1 == "B1",1],
       multiclass_dataset[multiclass_dataset$Sample_1 == "B1",2],
       pch = 1, col = "red")
points(multiclass_dataset[multiclass_dataset$Sample_1 == "C1",1],
       multiclass_dataset[multiclass_dataset$Sample_1 == "C1",2],
       pch = 1, col = "blue")

points(trial_simulation[trial_simulation$Sample_1 == "A1", 1],
       trial_simulation[trial_simulation$Sample_1 == "A1", 2],
       pch = 19, col = "black", cex = 2)
points(trial_simulation[trial_simulation$Sample_1 == "B1", 1],
       trial_simulation[trial_simulation$Sample_1 == "B1", 2],
       pch = 19, col = "red", cex = 2)
points(trial_simulation[trial_simulation$Sample_1 == "C1", 1],
       trial_simulation[trial_simulation$Sample_1 == "C1", 2],
       pch = 19, col = "blue", cex = 2)

plot(multiclass_dataset[multiclass_dataset$Sample_2 == "A2",1],
     multiclass_dataset[multiclass_dataset$Sample_2 == "A2",2],
     asp = 1, pch = 1,
     xlim = c(min(multiclass_dataset$x), max(multiclass_dataset$x)),
     ylim = c(min(multiclass_dataset$y), max(multiclass_dataset$y)))
points(multiclass_dataset[multiclass_dataset$Sample_2 == "B2",1],
       multiclass_dataset[multiclass_dataset$Sample_2 == "B2",2],
       pch = 1, col = "red")

points(trial_simulation[trial_simulation$Sample_2 == "A2", 1],
       trial_simulation[trial_simulation$Sample_2 == "A2", 2],
       pch = 19, col = "black", cex = 2)
points(trial_simulation[trial_simulation$Sample_2 == "B2", 1],
       trial_simulation[trial_simulation$Sample_2 == "B2", 2],
       pch = 19, col = "red", cex = 2)

#

}
\author{
Lloyd A. Courtenay
}
