% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericMC.R
\name{numericMC}
\alias{numericMC}
\title{Monte Carlo Simulations for Numeric Variables}
\usage{
numericMC(data_set, n_simulations, num_breaks = 0)
}
\arguments{
\item{data_set}{An matrix or array of numeric data to simulate}

\item{n_simulations}{An integer defining the number of simulations}

\item{num_breaks}{An optional parameter that can be used to fine-tune the calculation
of the probability distribution from which to sample from.}
}
\value{
simulated data
}
\description{
The present function can be used to simulate numerical (quantitative) data
}
\section{Details}{

This function uses the statistical chain rule to calculate multivariate probability
vectors associated with each categorical variable from which to sample from
}

\examples{

data(multiclass_dataset)
data_set <- as.matrix(multiclass_dataset[,c(1, 2)])

# augment data

augmented_data <- numericMC(data_set, 1000)

#

}
\author{
Lloyd A. Courtenay
}
