% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\name{sample_from_trace}
\alias{sample_from_trace}
\title{Sample from an MCMC Trace}
\usage{
sample_from_trace(traceObject, n_simulations)
}
\arguments{
\item{traceObject}{an MCMC_Trace object}

\item{n_simulations}{An integer defining the number of simulations to extract
from the MCMC_Trace}
}
\value{
Matrix of simulated values sampled from the trace
}
\description{
The present function can be used to sample values from an MCMC trace
}
\details{
This function first removes duplicate values from the MCMC's trace,
followed by a random sampling of the trace to produce a final matrix of new
simulated values.
}
\examples{

data("single_moon")

augmentation_trial <- MCMC(as.matrix(single_moon), 10000, step_size = 1)

trace <- sample_from_trace(augmentation_trial, 100)

plot(single_moon, asp = 1, pch = 19)
points(trace[,1], trace[,2], col = "red", pch  = 19)

#

}
\seealso{
\code{\link{MCMC}}, \code{\link{MCMC_Trace}}, \code{\link{burn_in}},
\code{\link{effective_sample_size}}, \code{\link{TOST}}, \code{\link{autocorrelation}},
\code{\link{trace_plot}}
}
\author{
Lloyd A. Courtenay
}
