% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\name{setTrace}
\alias{setTrace}
\title{Define a MCMC_Trace object}
\usage{
setTrace(traceObject, trace)
}
\arguments{
\item{traceObject}{an MCMC_Trace object}

\item{trace}{a numeric vector, matrix, or array}
}
\description{
The is function is to define an MCMC_Trace object.
}
\author{
Lloyd A. Courtenay
}
