% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentationMC_support.R
\docType{data}
\name{single_moon}
\alias{single_moon}
\title{Toy Single Moon Dataset}
\format{
A data frame with 500 rows and 2 variables (x, y)
}
\description{
A toy dataset containing x and y coordinates of a single multivariate moon
}
\author{
Lloyd A. Courtenay
}
\keyword{datasets}
