% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\name{TOST}
\alias{TOST}
\alias{TOST,MCMC_Trace-method}
\alias{TOST,matrix-method}
\alias{TOST,numeric-method}
\title{Two-One Sided Test (TOST) for Equivalence between Trace and Original Data}
\usage{
TOST(trace, original_data, dimension = NULL, robust = TRUE, plot = TRUE)

\S4method{TOST}{MCMC_Trace}(trace, original_data, dimension = NULL, robust = TRUE, plot = TRUE)

\S4method{TOST}{matrix}(trace, original_data, dimension = NULL, robust = TRUE, plot = TRUE)

\S4method{TOST}{numeric}(trace, original_data, robust = TRUE, plot = TRUE)
}
\arguments{
\item{trace}{either an MCMC_Trace object, a matrix of numeric values,
or a vector of numeric values.}

\item{original_data}{the original matrix}

\item{dimension}{an integer defining which dimension of the distribution should
be used for TOST calculations (in the case of multivariate data)}

\item{robust}{a boolean value defining whether the robust calculation of TOST
should be performed. Default is set to TRUE.}

\item{plot}{a boolean value defining whether a density plot should be produced
comparing the original and the augmented data}
}
\description{
Calculation of the statistical equivalence between the augmented data
sampled by the MCMC algorithm and the original data used to train the MCMC
}
\details{
This function can be used to compare two distributions to check for
their equivalence, and the magnitude of said equivalence. As recommended by Courtenay
and Gonzalez-Aguilera (2020), a useful way of testing the quality of synthetically
produced data is by using a Two-One Sided Test for Equivalence. From this perspective
the null hypothesis (pH_{0}) is that the two samples are different. Low p-Values
thus represent a high level of equivalence between the synthetically produced data
and the original data.

The way TOST is calculated depends on sample normality. If distributions are
Gaussian in nature, then a traditional TOST test using Welch's Test Statistic
can be used to calculate the magnitude of differences and similarities. Nevertheless,
if the distributions are not Gaussian, then a non-parametric test statistic should be
used. In this case we use the Yuen-Welch Trimmed test statistic, here referred to as
robust TOST (or rTOST)

The calculated absolute value is a normalised value, where both distribution 1
and distribution 2 are scaled to contain values between 0 and 1.

This test also returns (when p-Values are found to be under 0.3681) the probability
that the p-Value is capturing a false positive (Type I Statistical Error). The reported
probability uses a 0.5 prior, however is also accompanied by [0.7, 0.3] prior confidence
intervals
}
\section{Bibliography}{


Yuen, K.K.; Dixon, W.J. (1973) The Approximate Behaviour and Performance of the
Two-Sample Trimmed T. Biometrika. 60:369-374

Yuen, K.K. (1974) The Two-Sample Trimmed T for Unequal Population Variances. Biometrika.
61:165-170

Hauk, D.W.W.; Anderson, S. (1984) A New Statistical Procedure for Testing Equivalence
in Two-Group Comparative Bioavailability Trials. Journal of Pharmacokinetics and
Biopharmaceutics. 12:83-91

Schurimann, D.L. (1987) A comparison of the two one-sided test procedure and the
power approach for assessing the equivalence of average biovariability.
Journal of Pharmacokinetics and Biopharmaceutics. 15:657-680

Cohen, J. (1988) Statistical Power Analysis for Behavioural Sciences. New York: Routledge

Courtenay, L.A.; Gonzalez-Aguilera, D. (2020) Geometric Morphometric Data Augmentation using
Generative Computational Learning Algorithms, Applied Sciences.10:9133. DOI:
10.3390/app10249133
}

\examples{

data("single_moon")

augmentation_trial <- MCMC(as.matrix(single_moon), 10000, step_size = 1)

trace <- sample_from_trace(augmentation_trial, 100)

plot(single_moon, asp = 1, pch = 19)
points(trace[,1], trace[,2], col = "red", pch  = 19)

TOST(trace, single_moon, dimension = 1, robust = TRUE)

#

}
\seealso{
\code{\link{MCMC}}, \code{\link{MCMC_Trace}}, \code{\link{burn_in}},
\code{\link{sample_from_trace}},
\code{\link{effective_sample_size}}, \code{\link{autocorrelation}},
\code{\link{trace_plot}}
}
\author{
Lloyd A. Courtenay
}
