% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Trace.R
\name{trace_plot}
\alias{trace_plot}
\title{MCMC Trace Plots}
\usage{
trace_plot(trace, dimension = NULL, main = NULL)
}
\arguments{
\item{trace}{an MCMC_Trace object}

\item{dimension}{an integer defining which dimension of the distribution should
be visualised (if the trace is multivariate)}

\item{main}{an optional plot title}
}
\description{
Plot the Trace of a MCMC
}
\details{
Trace plots are a visual and useful guide to ensuring that the MCMC
has converged on the original distribution. Traces that do not converge will
appear to slowly move around the distribution, while traces that do converge
will fixate on where the original distribution is.
}
\examples{

data("single_moon")

trial_1 <- MCMC(as.matrix(single_moon[,1]), 10000, step_size = 1)
trial_2 <- MCMC(as.matrix(single_moon[,1]), 10000, step_size = 0.5)
trial_3 <- MCMC(as.matrix(single_moon[,1]), 10000, step_size = 0.1)


# example of convergence
trace_plot(trial_1, main = "Step Size = 1")

# example of poor convergence
trace_plot(trial_2, main = "Step Size = 0.5")

# example of no convergence
trace_plot(trial_3, main = "Step Size = 0.1")

#

}
\seealso{
\code{\link{MCMC}}, \code{\link{MCMC_Trace}}, \code{\link{burn_in}},
\code{\link{sample_from_trace}},
\code{\link{effective_sample_size}}, \code{\link{TOST}}, \code{\link{autocorrelation}}
}
\author{
Lloyd A. Courtenay
}
